Add warning for functions that are defined as exportable or public, but no declared in the proper header file.
What this means is that every function in a source (non-header) file
that is not static nor in an anonymous namespace should have a corresponding
declaration in a header file.  We expect the header file name to match
the source file we are checking (just replacing .cc/.cpp/.C/etc -> .h ).

There are some false positives introduced with this.  For example,
a main program isn't expected to have a corresponding header.
Then again, it probably shouldn't export any functions either. :-)


git-svn-id: http://cppclean.googlecode.com/svn/trunk@145 806ff5bb-693f-0410-b502-81bc3482ff28