Implement preliminary solution to #56 for exporting defines
This change modifies WarningHunter._determine_uses()._add_use() to take
in a full VariableDeclaration object instead of Type. This allows us to
check for the initial_value attribute of a VariableDeclaration object
which may contain a symbol from an included file (e.g. a #define). This
change also exports Define objects in order to capture their value and
check if they are being used.

To fully implement the issue rasied in #56, inline constructor member
initializer lists need to be supported.