Implement basic support for ctor initializer lists
This change adds basic support for initializing variables (initial_value
attribute of VariableDeclaration) inside ctor initializer lists. It only
supports one parameter for a given member. The ctor must also be inline
defined. In other words, a ctor with its definition outside the class is
not supported. Below is an example of what's supported:

```
class Foo
{
public:
  Foo() :
    foo(1)
  {}

private:
  int foo;
};
```

In this case, the `initial_value` of `foo` is set to `1`.

Whereas, the following is not supported, and will be ignored:

```
class Foo
{
public:
  Foo() :
    bar(1, 2, 3, 4)
  {}

private:
  Bar bar;
};
```

This ctor is defined outside the class, and thus not supported:

```
class Foo
{
public:
  Foo();

private:
  Bar bar;
};

Foo:Foo() :
  bar(1)
{}
```