Implement basic support for ctor initializer lists and exporting defines with values (#128)
* Implement preliminary solution to #56 for exporting defines

This change modifies WarningHunter._determine_uses()._add_use() to take
in a full VariableDeclaration object instead of Type. This allows us to
check for the initial_value attribute of a VariableDeclaration object
which may contain a symbol from an included file (e.g. a #define). This
change also exports Define objects in order to capture their value and
check if they are being used.

To fully implement the issue rasied in #56, inline constructor member
initializer lists need to be supported.

* Implement basic support for ctor initializer lists

This change adds basic support for initializing variables (initial_value
attribute of VariableDeclaration) inside ctor initializer lists. It only
supports one parameter for a given member. The ctor must also be inline
defined. In other words, a ctor with its definition outside the class is
not supported. Below is an example of what's supported:

```
class Foo
{
public:
  Foo() :
    foo(1)
  {}

private:
  int foo;
};
```

In this case, the `initial_value` of `foo` is set to `1`.

Whereas, the following is not supported, and will be ignored:

```
class Foo
{
public:
  Foo() :
    bar(1, 2, 3, 4)
  {}

private:
  Bar bar;
};
```

This ctor is defined outside the class, and thus not supported:

```
class Foo
{
public:
  Foo();

private:
  Bar bar;
};

Foo:Foo() :
  bar(1)
{}
```

* Add tests for ctor initializer lists

* Fix travis ci errors

* Fix using explicit python3 type

This fixes the broken build for Python 2.

* Fix inconsistency with usage of local variable